+++ R/removeFreq.Spectra2D.R
@@ -2,7 +2,6 @@
 #' removeFreq.Spectra2D
 #'
 #' @noRd
-#' @importFrom plyr is.formula
 #' @export
 #'
 removeFreq.Spectra2D <- function(spectra, rem.freq = NULL, remF2 = NULL, remF1 = NULL) {
@@ -13,7 +12,7 @@ removeFreq.Spectra2D <- function(spectra, rem.freq = NULL, remF2 = NULL, remF1 =
   # Subset data as requested
 
   if (!is.null(remF2)) { # F2 dimension: sorted F2 runs e.g. 0...10
-    if (!is.formula(remF2)) stop("remF2 must be a formula")
+    if (!inherits(remF2, "formula")) stop("remF2 must be a formula")
     limits <- .getLimits(spectra, "F2", remF2)
     toss <- !((spectra$F2 >= limits[1]) & (spectra$F2 <= limits[2]))
     for (i in 1:length(spectra$data)) spectra$data[[i]] <- spectra$data[[i]][, rev(toss), drop = FALSE]
@@ -21,7 +20,7 @@ removeFreq.Spectra2D <- function(spectra, rem.freq = NULL, remF2 = NULL, remF1 =
   }
 
   if (!is.null(remF1)) { # F1 dimension: sorted F1 runs e.g. 0...10 unsorted F1 runs e.g. 10...0
-    if (!is.formula(remF1)) stop("remF1 must be a formula")
+    if (!inherits(remF1, "formula")) stop("remF1 must be a formula")
     limits <- .getLimits(spectra, "F1", remF1)
     toss <- !((spectra$F1 >= limits[1]) & (spectra$F1 <= limits[2]))
     for (i in 1:length(spectra$data)) spectra$data[[i]] <- spectra$data[[i]][toss, , drop = FALSE]